# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Finalechat API",
    "version": "1.0.0",
    "summary": "The app where agents message the user.",
    "description": "Threads of messages and questions between agent sessions and one user, with long-polling, server-sent events and Web Push. Human-readable reference: https://www.finalechat.com/api/ . Agent quick start: https://www.finalechat.com/AGENTS.md .",
    "contact": {
      "url": "https://www.finalechat.com"
    }
  },
  "servers": [
    {
      "url": "https://www.finalechat.com/api/v1"
    },
    {
      "url": "https://api.finalechat.com/api/v1"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    },
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Create an account (app only, same-origin, rate limited)",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created; session cookie set.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  },
                  "required": [
                    "user"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Sign in (app only, same-origin, rate limited)",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed in; session cookie set.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  },
                  "required": [
                    "user"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/me": {
      "get": {
        "tags": [
          "Me"
        ],
        "summary": "Current user, counts and settings",
        "responses": {
          "200": {
            "description": "Identity.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Me"
        ],
        "summary": "Change display name or password (session only)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfilePatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  },
                  "required": [
                    "user"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/settings": {
      "get": {
        "tags": [
          "Me"
        ],
        "summary": "Read settings",
        "responses": {
          "200": {
            "description": "Settings.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "settings": {
                      "$ref": "#/components/schemas/Settings"
                    }
                  },
                  "required": [
                    "settings"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Me"
        ],
        "summary": "Update settings",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsPatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Full settings after the change.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "settings": {
                      "$ref": "#/components/schemas/Settings"
                    }
                  },
                  "required": [
                    "settings"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/counts": {
      "get": {
        "tags": [
          "Me"
        ],
        "summary": "Badge counts",
        "responses": {
          "200": {
            "description": "Counts.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "counts": {
                      "$ref": "#/components/schemas/Counts"
                    }
                  },
                  "required": [
                    "counts"
                  ]
                },
                "example": {
                  "counts": {
                    "pending_questions": 1,
                    "unread_threads": 3
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/events": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "Server-sent events stream",
        "description": "text/event-stream of `ready`, `thread.created`, `thread.updated`, `thread.deleted, thread.activity ({at, thread_id, activity}; no thread, no counts), question.dismissed, ping ({at} every 20 seconds)`, `message.created`, `question.created`, `question.answered`, `question.cancelled`, `question.expired`, `settings.updated` and `reconnect` events. Each data payload carries the full current objects (`thread`, `message`, `question`, `settings`) plus `counts` and `at`. A `: ping` comment is sent every 20 seconds.",
        "responses": {
          "200": {
            "description": "Event stream.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                },
                "example": "event: ready\ndata: {\"at\":\"2026-09-07T05:39:05.945123Z\",\"counts\":{\"pending_questions\":0,\"unread_threads\":3}}\n\nevent: message.created\ndata: {\"at\":\"…\",\"thread\":{…},\"message\":{…},\"counts\":{…}}\n\n"
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/tokens": {
      "get": {
        "tags": [
          "Tokens"
        ],
        "summary": "List active API tokens",
        "responses": {
          "200": {
            "description": "Tokens, newest first.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tokens": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/APIToken"
                      }
                    }
                  },
                  "required": [
                    "tokens"
                  ]
                },
                "example": {
                  "tokens": [
                    {
                      "id": "01a07a5a-be60-72d1-b7f0-8eb7c08f48b1",
                      "name": "laptop",
                      "prefix": "fc_TlaRtPqY",
                      "created_at": "2026-09-07T05:32:42.464521Z",
                      "last_used_at": "2026-09-07T05:38:23.601635Z"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tokens"
        ],
        "summary": "Create an API token (session only)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "maxLength": 120,
                    "default": "Agent token"
                  }
                }
              },
              "example": {
                "name": "laptop"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The token record and, once only, its secret.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "$ref": "#/components/schemas/APIToken"
                    },
                    "secret": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "token",
                    "secret"
                  ]
                },
                "example": {
                  "secret": "fc_TlaRtPqYwguWNmAtY9FXJLTEJQdiuQp7GYIvMWun",
                  "token": {
                    "id": "01a07a5a-be60-72d1-b7f0-8eb7c08f48b1",
                    "name": "laptop",
                    "prefix": "fc_TlaRtPqY",
                    "created_at": "2026-09-07T05:32:42.464521Z",
                    "last_used_at": null
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/tokens/{id}": {
      "delete": {
        "tags": [
          "Tokens"
        ],
        "summary": "Revoke an API token (session only)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "UUID."
            },
            "description": "Token id."
          }
        ],
        "responses": {
          "200": {
            "description": "Revoked.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/threads/{ref}/read": {
      "post": {
        "tags": [
          "Threads"
        ],
        "summary": "Mark the thread read",
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "description": "Thread UUID or `ext:<external_id>`.",
            "schema": {
              "type": "string"
            },
            "examples": {
              "uuid": {
                "value": "01a07a60-6dab-780a-bf4d-20ef15c0d7d7"
              },
              "external": {
                "value": "ext:claude-code:7f3a9c2e"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The thread.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "thread": {
                      "$ref": "#/components/schemas/Thread"
                    }
                  },
                  "required": [
                    "thread"
                  ]
                },
                "example": {
                  "thread": {
                    "id": "01a07a60-6dab-780a-bf4d-20ef15c0d7d7",
                    "external_id": "claude-code:7f3a9c2e",
                    "title": "finalechat",
                    "agent": "Claude Code",
                    "meta": {
                      "cwd": "/home/eric/finalechat"
                    },
                    "created_at": "2026-09-07T05:38:55.019126Z",
                    "updated_at": "2026-09-07T05:38:55.043799Z",
                    "last_activity_at": "2026-09-07T05:38:55.043799Z",
                    "last_read_at": "2026-09-07T05:38:55.019126Z",
                    "archived_at": null,
                    "muted": false,
                    "preview": "Staging looks good. Promote to production?",
                    "preview_sender": "question",
                    "unread_count": 1,
                    "pending_questions": 1,
                    "activity": null
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/messages/{id}": {
      "get": {
        "tags": [
          "Messages"
        ],
        "summary": "Get a message",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "UUID."
            },
            "description": "Message id."
          }
        ],
        "responses": {
          "200": {
            "description": "The message.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "$ref": "#/components/schemas/Message"
                    }
                  },
                  "required": [
                    "message"
                  ]
                },
                "example": {
                  "message": {
                    "id": "01a07a60-6db9-784b-bd24-7ba3d7d3dfa0",
                    "thread_id": "01a07a60-6dab-780a-bf4d-20ef15c0d7d7",
                    "sender": "agent",
                    "body": "Deployed **v1.2.0** to staging.\n\n- 42 tests passed\n- bundle 118 KB",
                    "format": "markdown",
                    "importance": "important",
                    "meta": {},
                    "created_at": "2026-09-07T05:38:55.033453Z",
                    "attachments": []
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/push/test": {
      "post": {
        "tags": [
          "Push"
        ],
        "summary": "Send a test notification to every subscribed device",
        "responses": {
          "202": {
            "description": "Queued.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "devices": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "ok",
                    "devices"
                  ]
                },
                "example": {
                  "ok": true,
                  "devices": 1
                }
              }
            }
          },
          "409": {
            "description": "No subscribed devices (`no_subscriptions`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Feature unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "fc_ token",
        "description": "API token created in the app under Settings → Agents. Usually supplied as the FINALECHAT_TOKEN environment variable."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "fc_session",
        "description": "Browser session set by POST /auth/login. State-changing requests must be same-origin."
      }
    },
    "schemas": {
...
      "SettingsPatchRequest": {
        "type": "object",
        "properties": {
          "notify_all_messages": {
            "type": "boolean"
          },
          "remote_mode": {
            "type": "boolean"
          }
        }
      },
      "ProfilePatchRequest": {
        "type": "object",
        "properties": {
          "display_name": {
            "type": "string",
            "maxLength": 80
          },
          "current_password": {
            "type": "string"
          },
          "new_password": {
            "type": "string",
            "minLength": 10,
            "maxLength": 1024
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 10,
            "maxLength": 1024
          },
          "display_name": {
            "type": "string"
          },
          "invite_code": {
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
    }
  }
}